refactor: refactor Themes, Footer and FormPageLayout components structure
Moved Footer and FormPageLayout main components to their own files and updated imports across the app for clarity and maintainability. Added new Global.css for shared styles and refactored Themes.css to remove global variables and improve theme separation. Updated ExcursionsPage styles to use new CSS variables. Ensured consistent imports and structure in related test files and main entry point.
