+++ src/css/Global.css
@@ -1,0 +1,82 @@
+/* ======================================== */
+/* 1. VARIABLES GLOBALES (NO DEPENDEN DEL TEMA) */
+/* ======================================== */
+:root {
+	/* Radios de borde para consistencia */
+	--border-radius-sm: 0.25rem; /* 4px */
+	--border-radius-md: 0.5rem; /* 8px */
+	--border-radius-lg: 0.75rem; /* 12px */
+}
+
+/* Estilos para botones deshabilitados de forma accesible. Se mantienen en el DOM y son enfocables, pero visualmente
+parecen inactivos. */
+.btn[aria-disabled="true"] {
+	opacity: 0.65; /* Reduce la opacidad para indicar visualmente que está inactivo */
+	cursor: not-allowed; /* Cambia el cursor para indicar que no es clickeable */
+}
+
+/* --- Menú de hamburguesa --- */
+.navbar-toggler-icon {
+	width: 1.3em;
+	height: 1.3em;
+}
+
+/* --- Offcanvas (Menú y Filtros) --- */
+.offcanvas-header {
+	font-weight: bolder;
+	padding: 1rem 1.25rem;
+}
+
+.offcanvas-body {
+	/* El padding por defecto de Bootstrap es 1rem. Lo eliminamos aquí para que los contenedores hijos
+	   puedan controlarlo. El Offcanvas de filtros, por ejemplo, tiene su propio padding interno. */
+	padding: 0;
+}
+
+/* --- Estilos Globales --- */
+/* Elimina transiciones por defecto para evitar parpadeos al cambiar de tema */
+.form-control,
+.page-link,
+.accordion-button,
+.accordion-body,
+body {
+	transition: none;
+}
+
+/* --- Formularios (Inputs, etc.) --- */
+/* Se elimina el box-shadow por defecto de Bootstrap para que el outline de :focus-visible sea el único indicador */
+.form-control:focus {
+	box-shadow: none;
+}
+
+/* Reseteamos la opacidad que Bootstrap aplica por defecto a los campos deshabilitados */
+.form-control:disabled:not(.form-control-plaintext),
+.form-control[readonly]:not(.form-control-plaintext) {
+	opacity: 1;
+}
+
+/* --- Offcanvas (Menú y Filtros) --- */
+#offcanvasNavbar .offcanvas-body .nav-link {
+	display: block;
+	width: 95%;
+	margin: 0 auto 0.6rem; /* Centrado horizontalmente y con margen inferior */
+	text-align: center;
+	padding: 0.6rem 1rem;
+	text-decoration: none;
+	border-radius: 0.375rem;
+	transition: background-color 0.2s ease-in-out, color 0.2s ease-in-out;
+}
+
+#offcanvasNavbar .offcanvas-body .logoutLink {
+	display: block;
+	width: 95%;
+	margin: 0 auto 0.6rem;
+}
+
+/* Estilos para los badges de dificultad (se aplica con una clase global) */
+.difficulty-badge {
+	font-size: 0.9rem;
+	padding: 0.3em 0.65em;
+	font-weight: 500;
+	letter-spacing: 0.5px;
+}
