+++ src/index.js
@@ -1,27 +1,26 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import { BrowserRouter } from 'react-router-dom';
-import Store from './store/store'
-import { Provider } from 'react-redux';
-import reportWebVitals from './reportWebVitals';
-
+import React from "react";
+import ReactDOM from "react-dom/client";
+import "./index.css";
+import "bootstrap/dist/css/bootstrap.css";
+import "./css/Global.css";
+import "./css/Themes.css";
+import App from "./App";
+import { BrowserRouter } from "react-router-dom";
+import Store from "./store/store";
+import { Provider } from "react-redux";
+import reportWebVitals from "./reportWebVitals";
 
 // react v19.1
 
-ReactDOM.createRoot(document.getElementById('root')).render(
-
-  <React.StrictMode>
-    <Provider store={Store}>
-        <BrowserRouter>
-          <App />
-        </BrowserRouter>
-    </Provider>
-  </React.StrictMode>
-
+ReactDOM.createRoot(document.getElementById("root")).render(
+	<React.StrictMode>
+		<Provider store={Store}>
+			<BrowserRouter>
+				<App />
+			</BrowserRouter>
+		</Provider>
+	</React.StrictMode>
 );
-  
 
 // If you want to start measuring performance in your app, pass a function
 // to log results (for example: reportWebVitals(console.log))
